extract `lib.user.blocking`
